Move Seo Product & BreadCrumbList into a shared location
